backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-slugify==8.0.1
marshmallow==3.20.1
anthropic==0.8.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
PyJWT==2.8.0
cryptography==41.0.7
